+++ canvas-component-tools/canvas-component-tools/commands/create.js
@@ -27,6 +27,7 @@ const component = {
   handle: '',
   folder: '',
   formatted: {
+    description: '',
     folder: '',
     lowerCase: '',
     pascalCase: '',
@@ -287,7 +288,7 @@ async function descriptionQuestion() {
       name: 'answer',
       prefix: symbols.description,
       result(answer) {
-        return answer.slice(-1) === '.' ? answer : `${answer}.`
+        return answer.trim().slice(-1) === '.' ? answer.trim() : `${answer.trim()}.`
       },
       type: 'input',
       validate(answer) {
@@ -566,6 +567,25 @@ function formatAnswers() {
         return $1.toUpperCase() + $2.toLowerCase()
       },
     )
+
+  /**
+   * Split description onto multiple lines to avoid eslint errors.
+   */
+  const words = component.description.split(' ')
+  let lineLength = 0
+
+  const description = words.map((word, index) => {
+    lineLength += index === 0 ? word.length : word.length + 1
+
+    if (lineLength > 77) {
+      lineLength = word.length
+      return `\n * ${word}`
+    }
+
+    return index === 0 ? word : ` ${word}`
+  })
+
+  component.formatted.description = description.join('')
 }
 
 /**
@@ -605,7 +625,7 @@ async function buildComponent() {
 
     if (component.template === 'limited-interactivity') {
       templateFilepath.liquid = `liquid-${component.folder}-${component.liquid}-limited-interactivity`
-      templateFilepath.vue = 'vue-limited-interactivity'
+      templateFilepath.vue = `vue-${component.folder}-limited-interactivity`
     }
 
     if (component.template === 'static') {