+++ canvas-component-tools/CHANGELOG.md
@@ -2,6 +2,17 @@
 
 Canvas Component Tools uses [semantic versioning](https://semver.org/).
 
+## 1.4.1 - 2023-07-03
+
+* Fixed global limited interactivity template causing compile errors
+
+## 1.4.0 - 2023-05-30
+
+* Updated `disabled_on` property to be disabled on all section groups
+* Fixed description not wrapping in Vue template causing eslint errors
+* Fixed description not being formatted correctly
+* Updated dependencies
+
 ## 1.3.0 - 2023-04-11
 
 * Added `disabled_on` property to async schema files to disable in header and footer section groups by default