+++ canvas-storybook-tools/CHANGELOG.md
@@ -2,11 +2,20 @@
 
 Canvas Storybook Tools uses [semantic versioning](https://semver.org/).
 
-## 2.0.0 - 2023-05-??
+## 2.0.0 - 2023-07-??
 
 * Updated Storybook to version 7
 * Updated dependencies
 * Removed commands and helpers
+
+## 1.15.0 - 2023-06-06
+
+* Updated dependencies
+* Fixed security issues
+
+## 1.14.0 - 2023-05-18
+
+* Replaced `glob-fs` with `@we-make-websites/file-sync`
 
 ## 1.13.0 - 2023-04-11